DCCU now correctly skips CRC's and clears padding frames
When used with an MP1 file that has CRC's (such as the ones generated by NCH Wavepad and NCH Switch), the CRC's are correctly skipped and removed.

When used with an MP1 file at 44.1 kHz, padding slots are cleared so that DCC-Studio and the DCC-175 PASC decoder can correctly handle the data.